+++ engine_src/shared/T1_mtlparser.h
@@ -4,6 +4,8 @@
 #include <stdint.h>
 #include <stddef.h>
 
+#include "T1_stdint.h"
+
 typedef struct ParsedMaterial {
     char name[T1_MATERIAL_NAME_CAP];
     char ambient_map[T1_MATERIAL_NAME_CAP];
@@ -11,39 +13,39 @@ typedef struct ParsedMaterial {
     char specular_map[T1_MATERIAL_NAME_CAP];
     char specular_exponent_map[T1_MATERIAL_NAME_CAP];
     char bump_or_normal_map[T1_MATERIAL_NAME_CAP];
-    float bump_map_intensity;
-    float specular_exponent;
-    float refraction;
-    float alpha;
-    float illum;
-    float roughness; // Pr in mtl files
-    float metallic; // Pm in mtl files
-    float sheen; // Ps in mtl files
-    float clearcoat; // Pc in mtl files
-    float clearcoat_roughness; // Pcr in mtl files
-    float anisotropy; // aniso in mtl files
-    float anisotropy_rotation; // anisor in mtl files
-    float ambient_rgb[3];
-    float diffuse_rgb[3];
-    float specular_rgb[3];
-    float emissive_rgb[3];
-    float T1_uv_scroll[2];
-    uint8_t use_base_mtl_flag;
+    f32 bump_map_intensity;
+    f32 specular_exponent;
+    f32 refraction;
+    f32 alpha;
+    f32 illum;
+    f32 roughness; // Pr in mtl files
+    f32 metallic; // Pm in mtl files
+    f32 sheen; // Ps in mtl files
+    f32 clearcoat; // Pc in mtl files
+    f32 clearcoat_roughness; // Pcr in mtl files
+    f32 anisotropy; // aniso in mtl files
+    f32 anisotropy_rotation; // anisor in mtl files
+    f32 ambient_rgb[3];
+    f32 diffuse_rgb[3];
+    f32 specular_rgb[3];
+    f32 emissive_rgb[3];
+    f32 T1_uv_scroll[2];
+    u8 use_base_mtl_flag;
 } ParsedMaterial;
 
 void mtlparser_init(
-    void * (* memset_func)(void *, int, size_t),
-    void * (* malloc_func)(size_t),
-    size_t (* strlcat)(char *, const char *, size_t));
+    void * (* memset_func)(void *, int, u64),
+    void * (* malloc_func)(u64),
+    u64 (* strlcat)(char *, const char *, u64));
 
 const char * mtlparser_get_last_error_msg(void);
 
 void mtlparser_parse(
     ParsedMaterial * recipient,
-    uint32_t * recipient_size,
-    const uint32_t recipient_cap,
+    u32 * recipient_size,
+    const u32 recipient_cap,
     const char * input,
-    uint8_t * good);
+    u8 * good);
 
 #endif // MTLPARSER_H
