+++ engine_src/shared/T1_profiler.h
@@ -1,13 +1,13 @@
 #ifndef T1_PROFILER_H
 #define T1_PROFILER_H
 
-#include <stdint.h>
+#include "T1_stdint.h"
 
 #if T1_PROFILER_ACTIVE == T1_ACTIVE
 
 void T1_profiler_init(
-    const uint64_t clock_frequency,
-    void * (* profiler_malloc_function)(size_t));
+    const u64 clock_frequency,
+    void * (* profiler_malloc_function)(u64));
 
 void T1_profiler_new_frame(void);
