+++ engine_src/shared/T1_text.h
@@ -11,56 +11,56 @@ extern "C" {
 extern T1TextFontSettings * T1_text_props;
 
 void T1_text_init(
-    void * (* arg_text_malloc_func)(size_t size),
+    void * (* arg_text_malloc_func)(u64 size),
     const char * raw_fontmetrics_file_contents,
-    const uint64_t raw_fontmetrics_file_size);
+    const u64 raw_fontmetrics_file_size);
 
 void T1_text_request_label_offset_around(
-    const int32_t with_id,
+    const s32 with_id,
     const char * text_to_draw,
-    const float mid_x_pixelspace,
-    const float mid_y_pixelspace,
-    const float z,
-    const float max_width);
+    const f32 mid_x_pixelspace,
+    const f32 mid_y_pixelspace,
+    const f32 z,
+    const f32 max_width);
 
 void T1_text_request_label_around_x_at_top_y(
-    const int32_t with_object_id,
+    const s32 with_object_id,
     const char * text_to_draw,
-    const float mid_x_pixelspace,
-    const float top_y_pixelspace,
-    const float z,
-    const float max_width);
+    const f32 mid_x_pixelspace,
+    const f32 top_y_pixelspace,
+    const f32 z,
+    const f32 max_width);
 
 void T1_text_request_label_around(
-    const int32_t with_object_id,
+    const s32 with_object_id,
     const char * text_to_draw,
-    const float mid_x_pixelspace,
-    const float mid_y_pixelspace,
-    const float z,
-    const float max_width);
+    const f32 mid_x_pixelspace,
+    const f32 mid_y_pixelspace,
+    const f32 z,
+    const f32 max_width);
 
 void T1_text_request_label_renderable(
-    const int32_t with_object_id,
+    const s32 with_object_id,
     const char * text_to_draw,
-    const float left_pixelspace,
-    const float top_y_pixelspace,
-    const float z,
-    const float tab_width,
-    const float max_width);
+    const f32 left_pixelspace,
+    const f32 top_y_pixelspace,
+    const f32 z,
+    const f32 tab_width,
+    const f32 max_width);
 
 void T1_text_request_label_leftx_toplinemidy(
-    const int32_t with_object_id,
+    const s32 with_object_id,
     const char * text_to_draw,
-    const float left_pixelspace,
-    const float topline_mid_y_pixelspace,
-    const float z,
-    const float max_width);
+    const f32 left_pixelspace,
+    const f32 topline_mid_y_pixelspace,
+    const f32 z,
+    const f32 max_width);
 
 void T1_text_request_debug_text(const char * text);
-void T1_text_request_fps(uint64_t elapsed_us);
+void T1_text_request_fps(u64 elapsed_us);
 
 void T1_text_request_top_touch_id(
-    int32_t top_touchable_id);
+    s32 top_touchable_id);
 
 #ifdef __cplusplus
 }
